requirements.txt
pytesseract
opencv-python
pyrebase4
flask
gunicorn